upgraded graphics

+++ include/graphics/gfx.h
@@ -8,7 +8,7 @@ typedef struct
   uint16_t width;
   uint16_t height;
   uint16_t pitch;
-  uint8_t *buffer;
+  uint32_t *buffer;
   uint8_t active;
 } gfx_display_t;
 
@@ -22,7 +22,7 @@ typedef struct
 
 typedef struct
 {
-  uint8_t color;
+  uint32_t color;
 } gfx_gc_t;
 
 gfx_display_t *gfx_open_display(void);
@@ -31,11 +31,11 @@ void gfx_close_display(gfx_display_t *display);
 gfx_window_t gfx_create_simple_window(gfx_display_t *display, int16_t x, int16_t y,
                                       uint16_t width, uint16_t height);
 
-gfx_gc_t gfx_create_gc(gfx_display_t *display, uint8_t color);
-void gfx_set_foreground(gfx_display_t *display, gfx_gc_t *gc, uint8_t color);
+gfx_gc_t gfx_create_gc(gfx_display_t *display, uint32_t color);
+void gfx_set_foreground(gfx_display_t *display, gfx_gc_t *gc, uint32_t color);
 void gfx_map_window(gfx_display_t *display, gfx_window_t *window);
 void gfx_unmap_window(gfx_display_t *display, gfx_window_t *window);
-void gfx_clear_window(gfx_display_t *display, gfx_window_t *window, uint8_t color);
+void gfx_clear_window(gfx_display_t *display, gfx_window_t *window, uint32_t color);
 void gfx_draw_point(gfx_display_t *display, gfx_window_t *window, gfx_gc_t *gc,
                     int16_t x, int16_t y);
 void gfx_draw_line(gfx_display_t *display, gfx_window_t *window, gfx_gc_t *gc,
